register templates in adaptivedialog
