+++ libraries/Microsoft.Bot.Builder.Dialogs.Adaptive/Generators/LanguageGeneratorManager.cs
@@ -35,6 +35,7 @@ namespace Microsoft.Bot.Builder.Dialogs.Adaptive.Generators
         {
             this.resourceExplorer = resourceExplorer;
             multilanguageResources = LGResourceLoader.GroupByLocale(resourceExplorer);
+            TemplatesMapping = new ConcurrentDictionary<(string resourceId, string locale), LanguageGeneration.Templates>();
 
             // load all LG resources
             foreach (var resource in this.resourceExplorer.GetResources("lg"))
@@ -44,7 +45,6 @@ namespace Microsoft.Bot.Builder.Dialogs.Adaptive.Generators
 
             // listen for resource changes
             this.resourceExplorer.Changed += ResourceExplorer_Changed;
-            RegisterTemplateFunctions();
         }
 
         /// <summary>
@@ -107,124 +107,6 @@ namespace Microsoft.Bot.Builder.Dialogs.Adaptive.Generators
         private TemplateEngineLanguageGenerator GetTemplateEngineLanguageGenerator(Resource resource)
         {
             return new TemplateEngineLanguageGenerator(resource, multilanguageResources);
-        }
-
-        private void RegisterTemplateFunctions()
-        {
-            var allTemplateNames = new List<string>();
-            TemplatesMapping.Values.ToList().ForEach(u => allTemplateNames.AddRange(u.AllTemplates.Select(t => t.Name)));
-            allTemplateNames = allTemplateNames.Distinct().ToList();
-
-            var prebuildFunctionNames = Expression.Functions.Values.Select(u => u.Type);
-            foreach (var templateName in allTemplateNames)
-            {
-                var functionType = prebuildFunctionNames.Contains(templateName) ? "lg." + templateName : templateName;
-                Expression.Functions.Add(functionType, new ExpressionEvaluator(
-                    functionType,
-                    (expression, state, options) =>
-                    {
-                        var currentLocale = GetCurrentLocale(state, options);
-
-                        if (state.TryGetValue(AdaptiveDialog.GeneratorIdKey, out var resourceId))
-                        {
-                            var (resourceName, _) = LGResourceLoader.ParseLGFileName(resourceId.ToString());
-
-                            var languagePolicy = GetLanguagePolicy(state);
-
-                            var fallbackLocales = GetFallbackLocales(languagePolicy, currentLocale);
-
-                            foreach (var fallbackLocale in fallbackLocales)
-                            {
-                                var templatesExist = TemplatesMapping.TryGetValue((resourceName, fallbackLocale.ToLowerInvariant()), out var templates);
-                                if (!templatesExist || !templates.Exists(u => u.Name == templateName))
-                                {
-                                    continue;
-                                }
-
-                                var (result, error) = EvaluateTemplate(templates, templateName, expression, state, options, currentLocale);
-                                if (error == null)
-                                {
-                                    return (result, null);
-                                }
-                            }
-                        }
-
-                        throw new Exception($"{templateName} does not have an evaluator");
-                    }));
-            }
-        }
-
-        private (object result, string error) EvaluateTemplate(LanguageGeneration.Templates templates, string templateName, Expression expression, IMemory state, Options options, string currentLocale)
-        {
-            var (args, error) = FunctionUtils.EvaluateChildren(expression, state, options);
-            if (error == null)
-            {
-                var parameters = templates.First(u => u.Name == templateName).Parameters;
-                var newScope = parameters.Zip(args, (k, v) => new { k, v })
-                    .ToDictionary(x => x.k, x => x.v);
-                var scope = new StackedMemory();
-                scope.Push(state);
-                scope.Push(new SimpleObjectMemory(newScope));
-                var lgOpt = new EvaluationOptions() { Locale = currentLocale, NullSubstitution = options.NullSubstitution };
-                var result = templates.Evaluate(templateName, scope, lgOpt);
-                return (result, error);
-            }
-
-            return (null, error);
-        }
-
-        private LanguagePolicy GetLanguagePolicy(IMemory memory)
-        {
-            var getLanguagePolicy = memory.TryGetValue(AdaptiveDialog.LanguagePolicy, out var lp);
-            LanguagePolicy languagePolicy;
-            if (!getLanguagePolicy)
-            {
-                languagePolicy = new LanguagePolicy();
-            }
-            else
-            {
-                languagePolicy = JObject.FromObject(lp).ToObject<LanguagePolicy>();
-            }
-
-            return languagePolicy;
-        }
-
-        private List<string> GetFallbackLocales(LanguagePolicy languagePolicy, string currentLocale)
-        {
-            var fallbackLocales = new List<string>();
-
-            if (languagePolicy.ContainsKey(currentLocale))
-            {
-                fallbackLocales.AddRange(languagePolicy[currentLocale]);
-            }
-
-            // append empty as fallback to end
-            if (currentLocale.Length != 0 && languagePolicy.ContainsKey(string.Empty))
-            {
-                fallbackLocales.AddRange(languagePolicy[string.Empty]);
-            }
-
-            if (fallbackLocales.Count == 0)
-            {
-                throw new InvalidOperationException($"No supported language found for {currentLocale}");
-            }
-
-            return fallbackLocales;
-        }
-
-        private string GetCurrentLocale(IMemory memory, Options options)
-        {
-            string currentLocale;
-            if (memory.TryGetValue("turn.locale", out var locale))
-            {
-                currentLocale = locale.ToString();
-            }
-            else
-            {
-                currentLocale = options.Locale;
-            }
-
-            return currentLocale;
         }
     }
 }
