+++ libraries/Microsoft.Bot.Builder.LanguageGeneration/Expander.cs
@@ -502,10 +502,11 @@ namespace Microsoft.Bot.Builder.LanguageGeneration
         private EvaluatorLookup CustomizedEvaluatorLookup(EvaluatorLookup baseLookup, bool isExpander)
         => (string name) =>
         {
-            // Standard functions have the highest priority.
-            if (ExpressionFunctions.StandardFunctions.ContainsKey(name))
+            var standardFunction = baseLookup(name);
+
+            if (standardFunction != null)
             {
-                return ExpressionFunctions.StandardFunctions[name];
+                return standardFunction;
             }
 
             if (name.StartsWith("lg.", StringComparison.Ordinal))
@@ -523,13 +524,6 @@ namespace Microsoft.Bot.Builder.LanguageGeneration
                 {
                     return new ExpressionEvaluator(name, FunctionUtils.Apply(this.TemplateEvaluator(name)), ReturnType.Object, this.ValidTemplateReference);
                 }
-            }
-
-            var standardFunction = baseLookup(name);
-
-            if (standardFunction != null)
-            {
-                return standardFunction;
             }
 
             const string template = "template";
