+++ tests/Microsoft.Bot.Builder.Dialogs.Adaptive.Templates.Tests/LGGeneratorTests.cs
@@ -25,6 +25,7 @@ namespace Microsoft.Bot.Builder.AI.LanguageGeneration.Tests
 {
     public class LGGeneratorTests
     {
+
         public LGGeneratorTests()
         {
             ComponentRegistration.Add(new DeclarativeComponentRegistration());
